添加抽卡、抽卡记录查询实现

+++ src/main/java/com/weird/mapper/PackageCardMapper.java
@@ -12,7 +12,7 @@ public interface PackageCardMapper {
 
     int insertSelective(PackageCardModel record);
 
-    PackageCardModel selectByPrimaryKey(Long pk);
+    PackageCardModel selectByPrimaryKey(int pk);
 
     int updateByPrimaryKeySelective(PackageCardModel record);
